+++ packages/furo-data/src/lib/FieldNode.js
@@ -142,6 +142,12 @@ export class FieldNode extends EventTreeNode {
     });
 
     // will be broadcasted by dataObject
+    this.addEventListener('set-pristine-request', () => {
+      this._pristine = true;
+      this._validationDisabled = false;
+    });
+
+    // will be broadcasted by dataObject
     this.addEventListener('new-data-injected', () => {
       this._pristine = true;
       this._validationDisabled = false;