+++ packages/furo-app/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.44](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/app@1.3.43...@furo/app@1.3.44) (2021-06-18)
+
+**Note:** Version bump only for package @furo/app
+
+
+
+
+
 ## [1.3.43](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/app@1.3.42...@furo/app@1.3.43) (2021-05-31)
 
 **Note:** Version bump only for package @furo/app