+++ packages/furo-experiments/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.3](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/experiments@1.5.2...@furo/experiments@1.5.3) (2021-06-18)
+
+**Note:** Version bump only for package @furo/experiments
+
+
+
+
+
 ## [1.5.2](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/experiments@1.5.1...@furo/experiments@1.5.2) (2021-05-31)
 
 **Note:** Version bump only for package @furo/experiments