+++ packages/furo-fbp/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.5.0](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/fbp@5.4.2...@furo/fbp@5.5.0) (2021-06-18)
+
+
+### Features
+
+* ref searcher init ([4915e89](https://github.com/theNorstroem/FuroBaseComponents/commit/4915e89197db71e34125ebe7ae20770a2658fe14))
+
+
+
+
+
 ## [5.4.2](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/fbp@5.4.1...@furo/fbp@5.4.2) (2021-05-31)
 
 **Note:** Version bump only for package @furo/fbp