+++ packages/furo-icon/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.37](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/icon@1.3.36...@furo/icon@1.3.37) (2021-06-18)
+
+**Note:** Version bump only for package @furo/icon
+
+
+
+
+
 ## [1.3.36](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/icon@1.3.35...@furo/icon@1.3.36) (2021-05-31)
 
 **Note:** Version bump only for package @furo/icon