+++ packages/furo-timing/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.72](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/timing@1.3.71...@furo/timing@1.3.72) (2021-06-18)
+
+**Note:** Version bump only for package @furo/timing
+
+
+
+
+
+## [1.3.71](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/timing@1.3.70...@furo/timing@1.3.71) (2021-06-14)
+
+**Note:** Version bump only for package @furo/timing
+
+
+
+
+
 ## [1.3.70](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/timing@1.3.69...@furo/timing@1.3.70) (2021-06-07)
 
 **Note:** Version bump only for package @furo/timing